+++ lib/ui/groups/GroupsEventDetailPanel.dart
@@ -148,7 +148,7 @@ class _GroupEventDetailsPanelState extends State<GroupEventDetailPanel> with Not
       child: Stack(
         alignment: Alignment.bottomCenter,
         children: <Widget>[
-          AppString.isStringNotEmpty(_event?.imageURL) ?  Positioned.fill(child:Image.network(widget.event?.imageURL, fit: BoxFit.cover,)) : Container(),
+          AppString.isStringNotEmpty(_event?.exploreImageURL) ?  Positioned.fill(child:Image.network(widget.event?.exploreImageURL, fit: BoxFit.cover,)) : Container(),
           CustomPaint(
             painter: TrianglePainter(painterColor: Styles().colors.fillColorSecondaryTransparent05, left: false),
             child: Container(